Improved firestore dependencies
