+++ lib/src/firestore/extensions/extensions.dart
@@ -1,3 +1,4 @@
 export 'fire_values.dart';
 export 'path.dart';
 export 'path_segment.dart';
+export '../firestore.dart';
