+++ lib/src/firestore/extensions/path_segment.dart
@@ -1,6 +1,4 @@
-import 'package:firefast/firefast_core.dart';
-
-import '../firestore.dart';
+import 'extensions.dart';
 
 extension PathSegmentFirestoreExtensions on PathSegment {
   FirestoreDocumentPath toFirestoreDocument() =>
