+++ lib/src/firestore/firestore.dart
@@ -1,3 +1,4 @@
 export 'adapters/adapters.dart';
 export 'extensions/extensions.dart';
 export 'models/models.dart';
+export '../core.dart';
