+++ lib/src/firestore/models/models.dart
@@ -2,3 +2,4 @@ export 'datasource.dart';
 export 'firestore_document.dart';
 export 'output.dart';
 export 'path.dart';
+export '../firestore.dart';
