+++ lib/src/firestore/models/output.dart
@@ -1,4 +1,4 @@
-import 'package:firefast/firefast_core.dart';
+import 'models.dart';
 
 class FirestoreDocumentOutput extends OperationOutputReader {
   FirestoreDocumentOutput({required super.fields});
