+++ lib/src/firestore/models/path.dart
@@ -1,4 +1,4 @@
-import 'package:firefast/firefast_core.dart';
+import 'models.dart';
 
 class FirestoreCollectionPath extends PathSegment {
   const FirestoreCollectionPath(super.collection, {super.parent});
